pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-demo"
version = "0.1.0"
description = "A conversational AI agent that connects IBM watsonx.ai models with StepZen's MCP server for GraphQL data access."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-phoenix>=10.13.2",
    "langchain-ibm>=0.3.12",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "openinference-instrumentation-langchain>=0.1.43",
    "opentelemetry-api>=1.34.1",
    "opentelemetry-instrumentation>=0.55b1",
    "opentelemetry-instrumentation-threading>=0.55b1",
    "opentelemetry-sdk>=1.34.1",
    "python-dotenv>=1.1.0",
]

[project.scripts]
start = "mcp_demo.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
